requirements.txt
pandas
openai
google-genai
anthropic
tqdm